[fix] add leeway to the leeway
The tests that test the leeway argument set it to 3600s (1 hour).
The expired token fixture is set as expired 1 hours ago.
This means that if the tests are running on a slightly slow machine then
they can fail sometimes. This adds a bit of leeway to the leeway
argument.