[feature] add flask permissions commands
Adds two flask commands for inspecting and managing permissions.

`flask permissions list` lists permissions currently used on routes and
those found in the database. It will return an exit status of 0 if no
update is required and an exit status of 1 if an update is required. It
can therefore be used in the deployment pipeline to ensure we don't have
any permissions that are not in migrations.

`flask permissions update` generates an alembic migration to insert /
update / delete any permissions that need to be.

Generated migration notes:
http://alembic.zzzcomputing.com/en/latest/api/autogenerate.html

using raw SQL as that's all that is allowed for execute operations
https://github.com/zzzeek/alembic/blob/master/alembic/autogenerate/render.py#L782