Proof of concept authenticated client
This client shows how we might implement an authenticated API client
that automatically handles refreshing the access token. It also provides
an interface for assuming an end user's identity and refreshing that
token as well.

The DirectIdentityAuthenticator uses the API user's credentials and uses that
to refresh that auth token when it expires.

The AssumedIdentityAuthenticator instead makes an authenticated request
against the new /api/v1/auth/assume-identity endpoint. In order to
handle token expiry on the API user's credentials this authenticated
request is made through the provided client which in turn is
authenticated with the DirectIdentityAuthenticator.

In the worst case there are two requests to the user service in order to
service a single API request.